Correccionde tabla estatus, implementacion de columna"ëditar" en productos, cambio de tipo de dato en precio int por double

+++ src/main/java/com/mycompany/notasremisionsanantonio/logica/DetalleNota.java
@@ -27,6 +27,6 @@ public class DetalleNota {
     }
 
     public int calcularSubtotal() {
-        return producto.getPrecio() * cantidad;
+        return (int) (producto.getPrecio() * cantidad);
     }
 }
